removing/replacing scripts
